refactor login-form and register to each have its own css module
reorganize folder structure to keep component files together
