+++ src/components/login-form/login-form.module.css
@@ -1,49 +1,49 @@
-.login-container {
+.container {
   background-color: rgba(82, 70, 83, 1);
   height: 100vh;
 }
 
-.login-form {
+.form {
   background-color: rgba(208, 66, 122, 1);
   padding: 5em 2em;
   border-radius: 15px;
 }
-.form-input{
+.input{
   background-color: rgba(0, 0, 0, 0.3);
   margin-bottom: 3%;
   border-style: none;
   padding-left: 5px;
 }
 
-.form-input::placeholder { /* Chrome, Firefox, Opera, Safari 10.1+ */
+.input::placeholder { /* Chrome, Firefox, Opera, Safari 10.1+ */
 color: rgba(255, 255, 255, 0.75);
 opacity: 1; /* Firefox */
 }
 
-.form-input:-ms-input-placeholder { /* Internet Explorer 10-11 */
+.input:-ms-input-placeholder { /* Internet Explorer 10-11 */
   color:  rgba(255, 255, 255, 0.75);
 }
 
-.form-input::-ms-input-placeholder { /* Microsoft Edge */
+.input::-ms-input-placeholder { /* Microsoft Edge */
   color:  rgba(255, 255, 255, 0.75);
 }
 
-.form-btn {
+.btn {
   background-color: rgba(0, 0, 0, 0.15);
   color:  rgba(255, 255, 255);
 
 }
 
-.form-btn:focus{
+.btn:focus{
   color:  rgba(255, 255, 255);
 }
 
-.form-btn:hover {
+.btn:hover {
   background-color: rgba(0, 0, 0, 0.25);
   color:  rgba(255, 255, 255);
 }
 
-.form-btn:active {
+.btn:active {
   background-color: rgba(0, 0, 0, 0.40);
   color:  rgba(255, 255, 255);
 }
