+++ src/components/register/register.module.css
@@ -1,0 +1,49 @@
+.container {
+  background-color: rgba(82, 70, 83, 1);
+  height: 100vh;
+}
+
+.form {
+  background-color: rgba(208, 66, 122, 1);
+  padding: 5em 2em;
+  border-radius: 15px;
+}
+.input{
+  background-color: rgba(0, 0, 0, 0.3);
+  margin-bottom: 3%;
+  border-style: none;
+  padding-left: 5px;
+}
+
+.input::placeholder { /* Chrome, Firefox, Opera, Safari 10.1+ */
+color: rgba(255, 255, 255, 0.75);
+opacity: 1; /* Firefox */
+}
+
+.input:-ms-input-placeholder { /* Internet Explorer 10-11 */
+  color:  rgba(255, 255, 255, 0.75);
+}
+
+.input::-ms-input-placeholder { /* Microsoft Edge */
+  color:  rgba(255, 255, 255, 0.75);
+}
+
+.btn {
+  background-color: rgba(0, 0, 0, 0.15);
+  color:  rgba(255, 255, 255);
+
+}
+
+.btn:focus{
+  color:  rgba(255, 255, 255);
+}
+
+.btn:hover {
+  background-color: rgba(0, 0, 0, 0.25);
+  color:  rgba(255, 255, 255);
+}
+
+.btn:active {
+  background-color: rgba(0, 0, 0, 0.40);
+  color:  rgba(255, 255, 255);
+}
