+++ src/index.js
@@ -11,8 +11,8 @@ import {createStore, applyMiddleware} from 'redux';
 import { BrowserRouter, Route, Switch} from 'react-router-dom';
 import reducers from './reducers/index';
 import ChangePassword from './components/change-password';
-import LoginForm from './components/login-form';
-import RegisterForm from './components/register';
+import LoginForm from './components/login-form/login-form.js';
+import RegisterForm from './components/register/register.js';
 import AddNewExhibit from './components/add-new-exhibit';
 import AddNewGallery  from './components/add-gallery';
 import ShowGallery from './components/show_gallery';
